Refactor: move CashflowSankey layout geometry to frontend (fixes #49)
